signature-module-standalone/requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
google-api-python-client==2.33.0
google-auth
google-auth-httplib2
google-auth-oauthlib
openpyxl==3.1.2
pandas==2.2.0
python-dotenv>=0.19.0
